refactor(v3): consolidate duplicated DER helpers into shared der.ts
The SEV and TDX slices were built in parallel and each carried its own copy of
the strict-DER primitives (~117 identical LOC): readTLV/expectTag/children,
OID/uint/time decoding, extension parsing, PEM decode, byte equality, tag
constants. One shared engine-neutral der.ts now owns them; the platform files
keep only their genuinely divergent parts (SEV: RSA-PSS algorithm params, name
maps, key usage; TDX: CN extraction, ECDSA wrappers, PCS helpers). Diffing the
copies surfaced one real difference: TDX's parseExtensions silently accepted a
non-BOOLEAN critical flag — the merged version keeps the SEV/Go-faithful
rejection. Net -129 LOC. Verified: 183/183 Node, byte-identical vs Go, 184/184
browser suite, engine-neutral guard green.
